+++ src/components/pages/Diagnosis/ducks/fetch-patient-diagnoses-detail.duck.js
