+++ src/components/pages/Diagnosis/selectors.js
