+++ src/components/pages/Diagnosis/table-columns.config.js
@@ -1,7 +1,7 @@
 import { getDDMMMYYYY } from '../../../utils/time-helpers.utils';
 
 export const columnsConfig = [
-  { key: 'problem', title: 'Problems / Diagnoses', width: '56%' },
+  { key: 'problem', title: 'Diagnoses', width: '56%' },
   { key: 'dateOfOnset', title: 'Date', transformer: getDDMMMYYYY, width: '25%' },
   { key: 'source', title: 'Source', width: '19%' },
   { key: 'sourceId', title: 'SourceID', width: 0, display: 'none' },
