+++ src/config/for-test.constants.js
@@ -1,6 +1,6 @@
 export const testConstants = {
   hostName: 'http://dev.ripple.foundation:8000',
-  token: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1Mjg0NDg5MzQsImlhdCI6MTUyODQ0ODMzNCwiaXNzIjoicWV3ZC5qd3QiLCJhcHBsaWNhdGlvbiI6InJpcHBsZS1jZHItb3BlbmVociIsInRpbWVvdXQiOjYwMCwicWV3ZCI6IjUwM2IzNzFkN2ExOTZmYWY5YmE3YjdiNDhjZDFlOWRlZGEzMTZhOTJmM2E0ZDMzMjMyOTVkNTFkZGE4MDYxYzg2YTg0MDg3OWRlYWEyNjZkMDg5NThkNTBiOTVhZjJjZDIxMzc4ZjE4YmEwOGMxNTc3NTc4NTU5ZGRmZTE2Y2JkNjM1MDNiN2JiMDYwMmQ0NjU0N2E4N2FmZjNiN2M2Y2ZhZWEzM2ZhZTQ3MDRlNmIxZmM0YzBiNGIyOTA0NTA1NyIsIm5oc051bWJlciI6Ijk5OTk5OTkwMDAiLCJlbWFpbCI6Iml2b3IuY294QHJpcHBsZS5mb3VuZGF0aW9uIiwicm9sZSI6InBoclVzZXIiLCJ1aWQiOiJlNGE5N2ZhYS1iZDcyLTQ2ODMtYTZlMS02NGM4YTc1YjlhYzkiLCJvcGVuaWQiOnsianRpIjoiZDMwMGM4YjctM2QzYS00ODc1LThkY2EtOWI1ZmJmYjhmMWI0IiwiZXhwIjoxNTI4NDQ4NjIzLCJuYmYiOjAsImlhdCI6MTUyODQ0ODMyMywiaXNzIjoiaHR0cHM6Ly9rZXljbG9hay5kZXYxLnNpZ25pbi5uaHMudWsvY2ljYXV0aC9yZWFsbXMvTkhTIiwiYXVkIjoibGVlZHMtaGVsbSIsInN1YiI6IjcxYmE1MGYyLWI0M2ItNGVhNC1iZTBmLWY1YjdhZWJhMjlhNSIsInR5cCI6IklEIiwiYXpwIjoibGVlZHMtaGVsbSIsImF1dGhfdGltZSI6MTUyODQ0ODMyMywic2Vzc2lvbl9zdGF0ZSI6ImU0YTk3ZmFhLWJkNzItNDY4My1hNmUxLTY0YzhhNzViOWFjOSIsImFjciI6IjEiLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJpdm9yLmNveEByaXBwbGUuZm91bmRhdGlvbiIsImVtYWlsIjoiaXZvci5jb3hAcmlwcGxlLmZvdW5kYXRpb24iLCJuaHNOdW1iZXIiOiI5OTk5OTk5MDAwIiwiaWRfdG9rZW4iOiJleUpoYkdjaU9pSlNVekkxTmlJc0luUjVjQ0lnT2lBaVNsZFVJaXdpYTJsa0lpQTZJQ0kzV0dsTmRIVkdTSE4wTVZkdFlVcHBTWGRKWVZaSlpsSk9helE1U3psVE5YbFdhMHRoTWtaWlpIQTBJbjAuZXlKcWRHa2lPaUprTXpBd1l6aGlOeTB6WkROaExUUTROelV0T0dSallTMDVZalZtWW1aaU9HWXhZalFpTENKbGVIQWlPakUxTWpnME5EZzJNak1zSW01aVppSTZNQ3dpYVdGMElqb3hOVEk0TkRRNE16SXpMQ0pwYzNNaU9pSm9kSFJ3Y3pvdkwydGxlV05zYjJGckxtUmxkakV1YzJsbmJtbHVMbTVvY3k1MWF5OWphV05oZFhSb0wzSmxZV3h0Y3k5T1NGTWlMQ0poZFdRaU9pSnNaV1ZrY3kxb1pXeHRJaXdpYzNWaUlqb2lOekZpWVRVd1pqSXRZalF6WWkwMFpXRTBMV0psTUdZdFpqVmlOMkZsWW1FeU9XRTFJaXdpZEhsd0lqb2lTVVFpTENKaGVuQWlPaUpzWldWa2N5MW9aV3h0SWl3aVlYVjBhRjkwYVcxbElqb3hOVEk0TkRRNE16SXpMQ0p6WlhOemFXOXVYM04wWVhSbElqb2laVFJoT1RkbVlXRXRZbVEzTWkwME5qZ3pMV0UyWlRFdE5qUmpPR0UzTldJNVlXTTVJaXdpWVdOeUlqb2lNU0lzSW5CeVpXWmxjbkpsWkY5MWMyVnlibUZ0WlNJNkltbDJiM0l1WTI5NFFISnBjSEJzWlM1bWIzVnVaR0YwYVc5dUlpd2laVzFoYVd3aU9pSnBkbTl5TG1OdmVFQnlhWEJ3YkdVdVptOTFibVJoZEdsdmJpSXNJbTVvYzA1MWJXSmxjaUk2SWprNU9UazVPVGt3TURBaWZRLkN4Mk90a21WTXJTNXJ5TTQ4aHdYRENrQkFSMkM4RDY3MTBQNl9hQUFVMm9KUzA0ZEdqTGdwM1BsX0dJbVJGNEhCaTlRSFJuTVJRNGIxZnBTY24zZzNOQmdqYUR3QkEta2V2dDZfWVhHTXdGSmszdGVST0paNE5BSjg2QWloLXB2THc5VlVLaVdsYWprSkY3Y3V5Ulc3U05EUmdYbEd3cE8wZU80XzV5UG1HOUZ3TDN2d1ZYeDhmQ1MySkdoZ3dGWUdNZEU4M0FPbXhQRFNWQ1R3d3A2VmJjWFg4bEhHYVVNSTk3eEpIQzlCR2duVXN2dmY1TlZWUHlwOWFrZzQxYkExcHg2b0VDa1Y5and0ekFhX2lsNVZXOGEyZC14LTY1YnRVZ2owODA5Y0lvRjhOMl9mNWRwdVJTT1hKQlBWUzBtTWpTbU5WdmFGWG14Z05uc3NCdDZTdyJ9fQ.DyK3c50BMZlD81j0mETMB_UNDQE5djzIxv5sJbyMI4I'
+  token: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOjE1MjkwNDk1MjMsImlhdCI6MTUyOTA0ODkyMywiaXNzIjoicWV3ZC5qd3QiLCJhcHBsaWNhdGlvbiI6InJpcHBsZS1jZHItb3BlbmVociIsInRpbWVvdXQiOjYwMCwicWV3ZCI6ImQzOTAxY2FjNzczMmJlZjQ3NjY5MmY3NDFlNzNiNjYxYzI3NDFmNDMzZGQ4YmVmMTA3NDZjZWUxMzExMjQzYTQyNjE0YjE5ZDdhYzU1NjQ4MmNmZjFkNDEyNzU0YmRlYWNhNWNmYWUzZDFiNTc5MWExOGNhMzVhNTYzN2MwOTcyNDc2M2YzZDQ2NTlkNGUwZGViZDE5MThmNGEzYzFhNzViNmMxZGEzOTIwYmRjYzY3M2NjNWNiYzgyN2QxNGRiZiIsIm5oc051bWJlciI6Ijk5OTk5OTkwMDAiLCJlbWFpbCI6Iml2b3IuY294QHJpcHBsZS5mb3VuZGF0aW9uIiwicm9sZSI6InBoclVzZXIiLCJ1aWQiOiJlYTY3NTU2Yy1mZTU3LTRjZjAtYmFhMy1mNjRlYzNjZjRiYzAiLCJvcGVuaWQiOnsianRpIjoiOWZlNjhiMDEtNWI2ZC00ZDNkLWEyOGYtMmI2OGZjZThmOGYwIiwiZXhwIjoxNTI5MDQ5MjE2LCJuYmYiOjAsImlhdCI6MTUyOTA0ODkxNiwiaXNzIjoiaHR0cHM6Ly9rZXljbG9hay5kZXYxLnNpZ25pbi5uaHMudWsvY2ljYXV0aC9yZWFsbXMvTkhTIiwiYXVkIjoibGVlZHMtaGVsbSIsInN1YiI6IjcxYmE1MGYyLWI0M2ItNGVhNC1iZTBmLWY1YjdhZWJhMjlhNSIsInR5cCI6IklEIiwiYXpwIjoibGVlZHMtaGVsbSIsImF1dGhfdGltZSI6MTUyOTA0ODkxNiwic2Vzc2lvbl9zdGF0ZSI6ImVhNjc1NTZjLWZlNTctNGNmMC1iYWEzLWY2NGVjM2NmNGJjMCIsImFjciI6IjEiLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJpdm9yLmNveEByaXBwbGUuZm91bmRhdGlvbiIsImVtYWlsIjoiaXZvci5jb3hAcmlwcGxlLmZvdW5kYXRpb24iLCJuaHNOdW1iZXIiOiI5OTk5OTk5MDAwIiwiaWRfdG9rZW4iOiJleUpoYkdjaU9pSlNVekkxTmlJc0luUjVjQ0lnT2lBaVNsZFVJaXdpYTJsa0lpQTZJQ0kzV0dsTmRIVkdTSE4wTVZkdFlVcHBTWGRKWVZaSlpsSk9helE1U3psVE5YbFdhMHRoTWtaWlpIQTBJbjAuZXlKcWRHa2lPaUk1Wm1VMk9HSXdNUzAxWWpaa0xUUmtNMlF0WVRJNFppMHlZalk0Wm1ObE9HWTRaakFpTENKbGVIQWlPakUxTWprd05Ea3lNVFlzSW01aVppSTZNQ3dpYVdGMElqb3hOVEk1TURRNE9URTJMQ0pwYzNNaU9pSm9kSFJ3Y3pvdkwydGxlV05zYjJGckxtUmxkakV1YzJsbmJtbHVMbTVvY3k1MWF5OWphV05oZFhSb0wzSmxZV3h0Y3k5T1NGTWlMQ0poZFdRaU9pSnNaV1ZrY3kxb1pXeHRJaXdpYzNWaUlqb2lOekZpWVRVd1pqSXRZalF6WWkwMFpXRTBMV0psTUdZdFpqVmlOMkZsWW1FeU9XRTFJaXdpZEhsd0lqb2lTVVFpTENKaGVuQWlPaUpzWldWa2N5MW9aV3h0SWl3aVlYVjBhRjkwYVcxbElqb3hOVEk1TURRNE9URTJMQ0p6WlhOemFXOXVYM04wWVhSbElqb2laV0UyTnpVMU5tTXRabVUxTnkwMFkyWXdMV0poWVRNdFpqWTBaV016WTJZMFltTXdJaXdpWVdOeUlqb2lNU0lzSW5CeVpXWmxjbkpsWkY5MWMyVnlibUZ0WlNJNkltbDJiM0l1WTI5NFFISnBjSEJzWlM1bWIzVnVaR0YwYVc5dUlpd2laVzFoYVd3aU9pSnBkbTl5TG1OdmVFQnlhWEJ3YkdVdVptOTFibVJoZEdsdmJpSXNJbTVvYzA1MWJXSmxjaUk2SWprNU9UazVPVGt3TURBaWZRLm1wSHdvM2pwU3ZTQW1aR085bjYzbDl5UV9oNFdsZ3FMYlRVRFVlNGtSVk93M0JfalgxRklEVTd2Qi1HQ2VFRDBybU9sUDZza1hhQ3UyNldXQ2lpVTNNWVY4Ml9pc1F3YVJkdmZVSS1tb0R5VFJIeXpjNm80WEN0SUZUWEhZWWtnT2Z4T0FhOFUtZVJWYkYtNVB3czZ1NXlUWEwxbWp4U2VJSEk4NVJNaGhyS2N1a3JhTXFBcy15aVhWNGM1THVBOEM4S0U2QVNid1NvNGNfTm9Gd0VuV3JsaUxMQVBTTGx1ZHhad0NCUVdHQTRHYXc0RFh0RHVOdktnQnJrX1hpLXgwQWtOUzMyV0pIVzN5dXNBV29MRHZ5a3ZtaXFZVi1xRWNuOFhjVFBhaHk1NUpHZkJWN21BY2xNMjZUb3doRTRMSmlvVVF3Z2Y1NEdOMUM3dVpyTm1CZyJ9fQ.CV8BVqMHd8m3GYzcFL-Ss2nr-jbeKQkfOugK3_8fg-g'
 };
 
 export const isDevMode = false;
